requirements.txt
djangorestframework
python-dotenv
djangorestframework-simplejwt